requirements.txt
streamlit
numpy
pandas
tensorflow
Pillow